# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/views/{viewId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "viewId",
          "in": "path",
          "required": true
        }
      ],
      "put": {
        "summary": "",
        "operationId": "view-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      },
      "delete": {
        "summary": "",
        "operationId": "view-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "meta"
        ]
      }
    },
...
    "/tables/{tableId}/forms": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "form-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            }
          }
        },
        "description": "",
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Form"
              }
            }
          }
        }
      }
    },
    "/forms/{formId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "formId",
          "in": "path",
          "required": true
        }
      ],
      "put": {
        "summary": "",
        "operationId": "form-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Form"
              }
            }
          }
        }
      },
      "get": {
        "summary": "",
        "operationId": "form-read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Form"
                }
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
...
    "/tables/{tableId}/galleries": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "gallery-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Gallery"
              }
            }
          }
        }
      }
    },
    "/galleries/{galleryId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "galleryId",
          "in": "path",
          "required": true
        }
      ],
      "put": {
        "summary": "",
        "operationId": "gallery-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Gallery"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "",
        "operationId": "gallery-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ]
      },
      "get": {
        "summary": "",
        "operationId": "gallery-read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gallery"
                }
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
...
    "/tables/{tableId}/views": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "views-list",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ViewList"
          }
        },
        "tags": [
          "Meta"
        ],
        "description": ""
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Sort": {
        "title": "Sort",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_model_id": {
            "type": "string"
          },
          "fk_column_id": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
...
      "Grid": {
        "title": "Grid",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          }
        },
        "description": ""
      },
      "Gallery": {
        "title": "Gallery",
        "type": "object",
        "description": "",
        "properties": {
          "fk_view_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "next_enabled": {
            "type": "boolean"
          },
          "prev_enabled": {
            "type": "boolean"
          },
          "cover_image_idx": {
            "type": "integer"
          },
          "cover_image": {
            "type": "string"
          },
          "restrict_types": {
            "type": "string"
          },
          "restrict_size": {
            "type": "string"
          },
          "restrict_number": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GalleryColumn"
            }
          },
          "fk_model_id": {
            "type": "string"
          },
          "fk_cover_image_col_id": {
            "type": "string"
          }
        }
      },
      "GalleryColumn": {
        "title": "GalleryColumn",
        "type": "object",
        "description": "",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "help": {
            "type": "string"
          },
          "fk_col_id": {
            "type": "string"
          },
          "fk_gallery_id": {
            "type": "string"
          }
        }
      },
...
      "KanbanColumn": {
        "title": "KanbanColumn",
        "type": "object",
        "description": "",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "help": {
            "type": "string"
          },
          "fk_column_id": {
            "type": "string"
          },
          "fk_kanban_id": {
            "type": "string"
          }
        }
      },
      "Kanban": {
        "title": "Kanban",
        "type": "object",
        "description": "",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "public": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KanbanColumn"
            }
          },
          "fk_model_id": {
            "type": "string"
          }
        }
      },
      "Form": {
        "title": "Form",
        "type": "object",
        "description": "",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "heading": {
            "type": "string"
          },
          "subheading": {
            "type": "string"
          },
          "sucess_msg": {
            "type": "string"
          },
          "redirect_url": {
            "type": "string"
          },
          "redirect_after_secs": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "banner_image_url": {
            "type": "string"
          },
          "logo_url": {
            "type": "string"
          },
          "submit_another_form": {
            "type": "boolean"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormColumn"
            }
          },
          "fk_model_id": {
            "type": "string"
          }
        }
      },
      "FormColumn": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "_cn": "first_name",
            "alias": "first_name",
            "fk_column_id": "cl_yvyhsl9u81tokc",
            "id": "fvc_8z1i7t8aswkqxx",
            "fk_view_id": "vw_s1pf4umdnikoyn",
            "uuid": null,
            "label": "dsdsds",
            "help": null,
            "required": false,
            "show": 1,
            "order": 1,
            "created_at": "2022-02-15 12:39:04",
            "updated_at": "2022-02-15 12:39:16",
            "description": "dsdsdsdsd"
          }
        },
        "properties": {
          "fk_column_id": {
            "type": "string",
            "minLength": 1
          },
          "id": {
            "type": "string",
            "minLength": 1
          },
          "fk_view_id": {
            "type": "string",
            "minLength": 1
          },
          "uuid": {},
          "label": {
            "type": "string",
            "minLength": 1
          },
          "help": {},
          "required": {
            "type": "boolean"
          },
          "show": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "created_at": {
            "type": "string",
            "minLength": 1
          },
          "updated_at": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Paginated": {
        "title": "Paginated",
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer"
          },
          "totalRows": {
            "type": "integer"
          },
          "sort": {
            "type": [
              "string",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Sort"
            }
          },
          "isFirstPage": {
            "type": "boolean"
          },
          "isLastPage": {
            "type": "boolean"
          },
          "page": {
            "type": "number"
          }
        }
      },
...
      "ViewList": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "bases": {
              "list": [
                {
                  "id": "string",
                  "project_id": "string",
                  "alias": "string",
                  "host": "string",
                  "port": 0,
                  "username": "string",
                  "password": "string",
                  "database": "string",
                  "url": "string",
                  "params": "string",
                  "type": "string",
                  "ssl": "string"
                }
              ],
              "pageInfo": {
                "pageSize": 0,
                "totalRows": 0,
                "sort": "string",
                "isFirstPage": true,
                "isLastPage": true
              }
            }
          }
        },
        "properties": {
          "views": {
            "type": "object",
            "required": [
              "list",
              "pageInfo"
            ],
            "properties": {
              "list": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Grid"
                  },
                  {
                    "$ref": "#/components/schemas/Form"
                  },
                  {
                    "$ref": "#/components/schemas/Kanban"
                  },
                  {
                    "$ref": "#/components/schemas/Gallery"
                  }
                ]
              },
              "pageInfo": {
                "$ref": "#/components/schemas/Paginated"
              }
            }
          }
        },
        "required": [
          "views"
        ],
        "title": ""
      },
...
    },
...
    "responses": {
...
      "ViewList": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ViewList"
            },
            "examples": {
              "example-1": {
                "value": {
                  "views": {
                    "list": [
                      {
                        "id": "string",
                        "title": "string",
                        "alias": "string",
                        "deleted": true,
                        "order": 0
                      }
                    ],
                    "pageInfo": {
                      "pageSize": 0,
                      "totalRows": 0,
                      "sort": "string",
                      "isFirstPage": true,
                      "isLastPage": true
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}